+++ ProjectJBv2/src/dao/ICouponsDao.java
@@ -23,7 +23,7 @@ public interface ICouponsDao {
 	 * @param couponID Receive a coupon id and delete from data base
 	 * @throws Exception Can throw an exception
 	 */
-	void delete(int couponID) throws Exception;
+	void delete(long couponID) throws Exception;
 
 	/**
 	 * @param c Receive a coupon and update
@@ -37,7 +37,7 @@ public interface ICouponsDao {
 	 *                  that company
 	 * @throws Exception Can throw an exception
 	 */
-	public void delete(int couponID, int companyID) throws Exception;
+	public void delete(long couponID, long companyID) throws Exception;
 
 	/**
 	 * @return This function retrn all coupons
@@ -50,7 +50,7 @@ public interface ICouponsDao {
 	 * @return This function return coupon by his id
 	 * @throws Exception Can throw an exception
 	 */
-	Coupon getOneCoupon(int couponID) throws Exception;
+	Coupon getOneCoupon(long couponID) throws Exception;
 
 	/**
 	 * @param customerId Receive a customer id
@@ -58,7 +58,7 @@ public interface ICouponsDao {
 	 *                   vs coupons)
 	 * @throws Exception Can throw an exception
 	 */
-	void addCouponPurchase(int customerId, int couponId) throws Exception;
+	void addCouponPurchase(long customerId, long couponId) throws Exception;
 
 	/**
 	 * @param customerId Receive a customer id
@@ -66,14 +66,14 @@ public interface ICouponsDao {
 	 *                   customer vs coupons)
 	 * @throws Exception Can throw an exception
 	 */
-	void deleteCouponPurchase(int customerId, int couponId) throws Exception;
+	void deleteCouponPurchase(long customerId, long couponId) throws Exception;
 
 	/**
 	 * @param companyID Receive a company id
 	 * @return This function return all coupons by his company id
 	 * @throws Exception Can throw an exception
 	 */
-	public ArrayList<Coupon> getCompanyCouponsByID(int companyID) throws Exception;
+	public ArrayList<Coupon> getCompanyCouponsByID(long companyID) throws Exception;
 
 	/**
 	 * @param maxPrice  Receive a max price
@@ -81,7 +81,7 @@ public interface ICouponsDao {
 	 * @return This function return all coupons by his company id and max price
 	 * @throws Exception Can throw an exception
 	 */
-	public ArrayList<Coupon> getCompanyCouponsByMaxPrice(double maxPrice, int companyID) throws Exception;
+	public ArrayList<Coupon> getCompanyCouponsByMaxPrice(double maxPrice, long companyID) throws Exception;
 
 	/**
 	 * @param category  Receive a category
@@ -89,7 +89,7 @@ public interface ICouponsDao {
 	 * @return This function return all coupons by his company id and his category
 	 * @throws Exception Can throw an exception
 	 */
-	public ArrayList<Coupon> getCompanyCouponsByCategory(Category category, int companyID) throws Exception;
+	public ArrayList<Coupon> getCompanyCouponsByCategory(Category category, long companyID) throws Exception;
 
 	/**
 	 * @return This function return all id for expired coupons
